add email option, bring files on website up to match with local version

+++ admin/includes/view_all_comments.php
@@ -47,7 +47,7 @@
       echo "<td>{$comment_date}</td>";
       echo "<td><a href='comments.php?approve={$comment_id}'>Approve</a></td>";
       echo "<td><a href='comments.php?decline={$comment_id}'>Decline</a></td>";
-      echo "<td><a href='comments.php?delete={$comment_id}'>Delete</a></td>";
+      echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete?'); \" href='comments.php?delete={$comment_id}'>Delete</a></td>";
       echo "</tr>";
     }
     confirm($select_comments);
